src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --resolver=backtracking pyproject.toml
#
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via jsonschema
bcrypt==4.0.1
    # via Habiton-Backend (pyproject.toml)
certifi==2022.12.7
    # via
    #   django-healthchecks
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
cryptography==39.0.0
    # via pyjwt
django>=4.1.7
    # via
    #   Habiton-Backend (pyproject.toml)
    #   django-axes
    #   django-filter
    #   django-healthchecks
    #   django-storages
    #   djangorestframework
    #   drf-jwt
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-axes==5.40.1
    # via Habiton-Backend (pyproject.toml)
django-behaviors==0.5.1
    # via Habiton-Backend (pyproject.toml)
django-cors-headers==3.14.0
    # via Habiton-Backend (pyproject.toml)
django-debug-toolbar==3.8.1
    # via Habiton-Backend (pyproject.toml)
django-environ==0.9.0
    # via Habiton-Backend (pyproject.toml)
django-filter==22.1
    # via Habiton-Backend (pyproject.toml)
django-healthchecks==1.5.0
    # via Habiton-Backend (pyproject.toml)
django-ipware==4.0.2
    # via
    #   Habiton-Backend (pyproject.toml)
    #   django-axes
django-split-settings==1.2.0
    # via Habiton-Backend (pyproject.toml)
django-storages==1.13.2
    # via Habiton-Backend (pyproject.toml)
djangorestframework==3.14.0
    # via
    #   Habiton-Backend (pyproject.toml)
    #   drf-jwt
    #   drf-spectacular
djangorestframework-camel-case==1.3.0
    # via Habiton-Backend (pyproject.toml)
drf-jwt==1.19.2
    # via Habiton-Backend (pyproject.toml)
drf-spectacular[sidecar]==0.25.1
    # via Habiton-Backend (pyproject.toml)
drf-spectacular-sidecar==2022.12.1
    # via drf-spectacular
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.3
    # via drf-spectacular
pillow==9.4.0
    # via Habiton-Backend (pyproject.toml)
psycopg2-binary==2.9.5
    # via Habiton-Backend (pyproject.toml)
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via drf-jwt
pyrsistent==0.19.3
    # via jsonschema
pytz==2022.7.1
    # via
    #   django
    #   djangorestframework
pyyaml==6.0
    # via drf-spectacular
redis==4.4.2
    # via Habiton-Backend (pyproject.toml)
requests==2.28.2
    # via django-healthchecks
sentry-sdk==1.14.0
    # via Habiton-Backend (pyproject.toml)
sqlparse==0.4.3
    # via django
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.14
    # via
    #   requests
    #   sentry-sdk
whitenoise==6.3.0
    # via Habiton-Backend (pyproject.toml)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
